Fixed some build-time errors and warnings showing up with MSVC (Visual C++ 2010 Express). Fixed a crash on start-up of the unittest binary (though one test still fails).
git-svn-id: https://openmalaria.googlecode.com/svn/trunk@1286 aa1ca794-fd11-11dd-ac65-f74f249cfaac

+++ model/Population.cpp
@@ -356,7 +356,7 @@ void Population::implementIntervention (int time)
     }
     
     if (interv->getInsertR_0Case().present()){
-	int i = std::floor (random::uniform_01() * populationSize);	// pick a human
+	int i = (int)std::floor (random::uniform_01() * populationSize);	// pick a human
 	HumanIter it = population.begin();
 	while (i > 0){	// find human (can't use population[i])
 	    ++it;
